Release NDSL 2026.07.00

+++ ndsl/monitor/zarr_monitor.py
@@ -7,9 +7,9 @@ import cftime
 import xarray as xr
 
 import ndsl.constants as constants
+from ndsl import ndsl_log
 from ndsl.comm import Comm
 from ndsl.comm.partitioner import Partitioner, subtile_slice
-from ndsl.logging import ndsl_log
 from ndsl.monitor.convert import to_numpy
 from ndsl.optional_imports import cupy, zarr
 from ndsl.utils import list_by_dims
@@ -165,7 +165,7 @@ class _ZarrVariableWriter:
         self.array = self.group.create_dataset(
             self.name,
             shape=self._prepend_shape + tile_shape,
-            dtype=quantity.data.dtype,
+            dtype=quantity.dtype,
             chunks=chunks,
             fill_value=None,
         )
